+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+### 1.1.1
++ [x] Server: improved alerts for `FORK`, `FORWARD-FORK`
++ [x] Server: added alert for `NO PEERS` state
++ [x] Server: node info now stored into local object every time interval defined in `consfig.nodeInfoCollectInterval`
++ [x] Server: added config property `config.blockSpeedDistance` with default value `10` to calculate blockchain speed
++ [x] Server: information about mina now return from stored object
++ [x] Server: added request for node health, request return empty array if `OK` or filled array with critical states
+
++ [x] Client: added reaction for node health request
++ [x] Client: Node health now is displaying in sync status block
+
 ### 1.0.5 -> 1.1.0
 + [x] Server: fixed uptime sender for using `publicKeyDelegators`
 + [x] Server: fixed alerter module for forward for when unvalidated length is 0